pyproject.toml
[tool.poetry]
name = "django-pgviews-redux"
version = "0.13.0"
description = "Create and manage Postgres SQL Views in Django"
authors = ["Mikuláš Poul <mikulas.poul@xelix.com>"]
readme = "README.md"
packages = [{include = "django_pgviews"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
]
include = ["UNLICENSE"]
repository = "https://github.com/xelixdev/django-pgviews-redux"
keywords = ["django", "views", "materialized views", "postgres"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    # https://github.com/charliermarsh/ruff#pyflakes-f
    "F",
    # https://github.com/charliermarsh/ruff#pycodestyle-e-w
    "E",
    "W",
    # https://github.com/charliermarsh/ruff#isort-i
    "I",
    # https://github.com/charliermarsh/ruff#pep8-naming-n
    "N",
    # https://github.com/charliermarsh/ruff#pyupgrade-up
    "UP",
    # https://github.com/charliermarsh/ruff#flake8-bugbear-b
    "B",
    # https://github.com/charliermarsh/ruff#flake8-comprehensions-c4
    "C4",
    # https://github.com/charliermarsh/ruff#flake8-debugger-t10
    "T10",
    # https://github.com/charliermarsh/ruff#flake8-pie-pie
    "PIE",
    # https://github.com/charliermarsh/ruff#flake8-return-ret
    "RET",
    # https://github.com/charliermarsh/ruff#flake8-simplify-sim
    "SIM",
]

# Never enforce...
ignore = [
    "E501",   # line length violations
    "SIM108", # use-ternary-operator
    "RET505", # superfluous-else-return
    "RET506", # superfluous-else-raise
    "RET507", # superfluous-else-continue
    "RET508", # superfluous-else-break
    "B027",   # empty-method-without-abstract-decorator
]

[tool.pyrefly]
project-includes = [
    "django_pgviews/*.py*",
]

[dependency-groups]
dev = [
    "pyrefly (>=0.46.0,<0.47.0)",
    "django-stubs (>=5.2.8,<6.0.0)",
    "psycopg (>=3.3.2,<4.0.0)"
]
